Change of contracted session type representation
Contracted session types now carry a non-empty list of variables in mu
binders rather than a set. This ensures that the conversion from regular
to contracted session types is reversible; this property is proven in
Conv.v. Note, however, that the conversion in the opposite direction is
not reversible if the source type has a component of the form

  mu [X1, ..., Xn]. mu [Y1, ..., Ym]. S

This change of representation necessitates the reformulation of a bunch
of related theorems.

Conv.v also currently has a proof for

  wellformed S -> wellformed (conv S)

where S is a regular session type and (conv S) is the corresponding
contracted session type.

In order for the conversion to become useful, we still need

  S ~ T <-> (conv S) ~' (conv T)

where ~ and ~' denote equivalence.