Remove compacted session types
At this point, I believe that compacted session types don't result in
significantly simpler proofs, but they add a lot of complexity because
the conversion between regular and compacted session types has to be
proven correct, which turns out to be quite laborious.

This change will have to be re-evaluated depending on how bad the proofs
without contracted session types turn out to be.