Fix substall; rename to cs
substall used to have the wrong behavior for mu constructors:

    substall (mu X S) = subst X (mu X S) (substall S)

The correct definition is:

    substall (mu X S) = substall (subst X (mu X S) S)

I thought that this makes no difference, but it does because
substitutions don't in general commute, even under the assumptions we're
able to make in this case.

Unfortunately, this means that substall cannot be defined as a simple
fixpoint any more. We could define it in terms of wellformed recursion,
but that would require the argument session type to be wellformed, which
becomes ugly quickly. Thus, we opt to define it as a deterministic relation
instead, hoping that the proofs won't turn out too bad.

For now, we delete a couple of proofs which should not be required going
forward.

We use the occasion to rename substall to cs.