Better build system
coq_makefile apparently doesn't handle dependencies in subdirectories
too well, which manifested in a bug where the dependencies of
tlc/src/LibAxiom.v would not be built before that file (and the build
would thus fail).

To solve this issue, we now provide a handwritten Makefile which
delegates to the auto-generated Makefile.coq and tlc/src/Makefile. This
is a bit of an ugly hack because we have to duplicate each command from
these files in our new Makefile.

The new Makefile differs in one important aspect from the previous one:
`clean` will not recurse into the TLC directory, but the new `mrproper`
will.