Switch to Coq 8.5
This mostly means adjustments to the build system. Mainly, we don't use
the TLC Makefile at all and just add the TLC files to our auto-generated
one. This is less elegant, but a lot easier, and it means that all the
targets added in 8.5 (quick, checkproofs, vio2vo) work out of the box.