Trace languages and some basic facts about them
The developments regarding trace languages follow the paper, with the
crucial distinction that here, a trace language is effectively an
infinite syntax tree that trivially generates the sets which are usually
referred to as trace languages.

Trace languages are generated from contractive session types by the `tl`
function. This currently relies on two axioms because I could not figure
out how to convince Coq that the mixed recursion-corecursion employed in
that function is valid.