New build system; TLC update
The new build system (as cursorily described in the new README.md)
finally allows for separate compilation of TLC and this project, at the
expense of some more manual fiddling during the initial setup.

To enable the new build system, the folder structure has changed: This
project's source files are now in src/, while TLC is in lib/tlc/.

While we're at it, we update TLC to the latest available snapshot,
dropping Coq 8.4 compatibility (but this project has required 8.5 for
some time anyway).