Switch to new def. of wellformedness
We make the new definition of wellformedness introduced recently the
default, though we also keep the old definition (and a proof of its
equivalence) around.

The main benefit of this is a lot of deleted code related to the old
definition, including a rather nasty proof that substitution preserves
wellformedness. This property, and some others, are now proven more
easily. Besides, the new definition allows for simpler automation.