Remove beq_var and beq_msg; cleanup Env
This commit replaces all remaining uses of beq_var by eq_Var_dec, thus
rendering the former obsolete. It alse removes beq_Sty and beq_msg,
which we never used anyway.

To accomodate this change, a lot of proofs needed to be changed
slightly. The only bigger change is in WfAlt, where we introduce some
limited automation for Ok/Checked inversion in order to make the proofs
significantly simpler.

We also delete large portions of Env.v because they are not currently
used.